api-endpoint (1)/requirements.txt
# For FastAPI
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# For Django REST Framework
Django==5.0.1
djangorestframework==3.14.0
